--- a/.travis.yml
+++ b/.travis.yml
@@ -8,6 +8,14 @@ os:
   - linux
   - osx
 
+matrix:
+  include:
+    name: "Benchmarks"
+    language: go
+    go: stable
+    os: linux
+      - "bash ./.travis/benchmark.sh"
+
 script:
   - "go build"
   - "go test -v"