--- a/.travis.yml
+++ b/.travis.yml
@@ -14,6 +14,7 @@ matrix:
     language: go
     go: stable
     os: linux
+    script:
       - "bash ./.travis/benchmark.sh"
 
 script: